% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{getGeoTables}
\alias{getGeoTables}
\title{Load geographical data for Genetic Report Cards map plotting using the grcMAlaria package}
\usage{
getGeoTables()
}
\value{
A list containing ...
}
\description{
Load geographical data for Genetic Report Cards map plotting using the grcMAlaria package
}
\examples{
#TBD
}
